split the header by purpose
