+++ extension/resource_manager/include/resource_manager/stop/residual_norm.hpp
@@ -34,16 +34,19 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define GKO_PUBLIC_EXT_RESOURCE_MANAGER_STOP_RESIDUAL_NORM_HPP_
 
 
-#include <type_traits>
+#include <ginkgo/core/stop/residual_norm.hpp>
 
 
-#include "resource_manager/base/element_types.hpp"
+#include "resource_manager/base/generic_constructor.hpp"
 #include "resource_manager/base/helper.hpp"
 #include "resource_manager/base/macro_helper.hpp"
 #include "resource_manager/base/rapidjson_helper.hpp"
 #include "resource_manager/base/resource_manager.hpp"
-#include "resource_manager/base/type_list.hpp"
-
+#include "resource_manager/base/type_default.hpp"
+#include "resource_manager/base/type_pack.hpp"
+#include "resource_manager/base/type_resolving.hpp"
+#include "resource_manager/base/type_string.hpp"
+#include "resource_manager/base/types.hpp"
 
 namespace gko {
 namespace extension {
