+++ extension/resource_manager/test/stop/iteration.cpp
@@ -30,23 +30,14 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#include <thread>
-#include <type_traits>
-
-
-#include <ginkgo/core/base/executor.hpp>
-
-
-#if defined(__unix__) || defined(__APPLE__)
-#include <utmpx.h>
-#endif
+#include <memory>
 
 
 #include <gtest/gtest.h>
 #include <rapidjson/document.h>
 
 
-#include <ginkgo/core/base/exception.hpp>
+#include <ginkgo/core/stop/iteration.hpp>
 
 
 #include "resource_manager/resource_manager.hpp"
